cleans up classes object to be more sane and usable
