+++ src/characterData/classes/dwarf.ts
@@ -1,0 +1,25 @@
+export const dwarf = {
+  name: "Dwarf",
+  requirements: "Minimum CON 9",
+  primeRequisite: "STR",
+  hitDice: "1d8",
+  maximumLevel: "12",
+  allowedArmor: "Any, including shields",
+  allowedWeapons: "Small or normal sized (see below)",
+  languages: "Alignment language, Common, dwarvish, gnome, goblin, kobold",
+  saves: {
+    poison: 10,
+    wands: 11,
+    stone: 12,
+    breath: 13,
+    magic: 14
+  },
+  equipment: "see Fighter equipment",
+  abilities: [
+    "Combat: Dwarves may use any type of armour, however, due to their short height, may only use small or normal sized weapons. This means that dwarves cannot use longbows or two-handed swords.",
+    "Detect Construction Tricks: From their experience underground, dwarves have a 2-in-6 chance of detecting false walls, hidden construction, or sloping passages. A dwarf must be actively searching for these abilities to function.",
+    "Detect Traps: Due to their expertise with construction, dwarves have a 2-in-6 chance of detecting non-magical traps, when searching (see Searching in B/X Essentials: Core Rules).",
+    "Infravision: Due to their habit of living underground in great mines, dwarves have the ability to see in the dark with infravision up to 60’ (see Light, Vision, and Visibility in B/X Essentials: Core Rules).",
+    "Listening at Doors: Dwarves have a 2-in-6 chance of hearing noises (see Doors, B/X Essentials: Core Rules)."
+  ]
+};
