+++ src/utils/getClassPrimeRequisites.ts
@@ -1,27 +1,11 @@
-import {
-  cleric,
-  dwarf,
-  elf,
-  fighter,
-  halfling,
-  magicUser,
-  thief
-} from "../characterData/classes";
+import { characterClasses } from "../characterData/classes";
 
-const classPrimeRequisites: {
-  [key: string]: string;
-} = {
-  cleric: cleric.primeRequisite,
-  fighter: fighter.primeRequisite,
-  magicUser: magicUser.primeRequisite,
-  thief: thief.primeRequisite,
-  dwarf: dwarf.primeRequisite,
-  elf: elf.primeRequisite,
-  halfling: halfling.primeRequisite
-};
-
-export const getClassPrimeRequisites = (classOption: string) => {
-  const primeRequisites = classPrimeRequisites[classOption];
+/**
+ * Utility for getting a class' prime requisites.
+ * @param classOptionKey
+ */
+export const getClassPrimeRequisites = (classOptionKey: string) => {
+  const primeRequisites = characterClasses[classOptionKey].primeRequisite;
 
   return primeRequisites;
 };
